implment soft-delete functionality and add tests for it

+++ src/core/Wemogy.Infrastructure.Database.Core/Repositories/DatabaseRepository`1.Delete.cs
@@ -1,7 +1,7 @@
 using System;
+using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
-using Wemogy.Core.Errors.Exceptions;
 using Wemogy.Infrastructure.Database.Core.Abstractions;
 using Wemogy.Infrastructure.Database.Core.Errors;
 
@@ -12,24 +12,54 @@ public partial class DatabaseRepository<TEntity>
 {
     public async Task DeleteAsync(string id)
     {
-        var isDeleted = await _database.DeleteAsync(x => id == x.Id.ToString());
-        if (isDeleted == 1)
+        if (!SoftDeleteState.IsEnabled)
         {
+            var isDeleted = await _database.DeleteAsync(x => id == x.Id.ToString());
+            if (isDeleted == 1)
+            {
+                return;
+            }
+
+            throw DatabaseError.EntityNotFound(id);
+        }
+
+        await UpdateAsync(
+            id,
+            SoftDelete);
+    }
+
+    public async Task DeleteAsync(string id, string partitionKey)
+    {
+        if (!SoftDeleteState.IsEnabled)
+        {
+            await _database.DeleteAsync(
+                id,
+                partitionKey);
             return;
         }
 
-        throw DatabaseError.EntityNotFound(id);
-    }
-
-    public Task DeleteAsync(string id, string partitionKey)
-    {
-        return _database.DeleteAsync(
+        await UpdateAsync(
             id,
-            partitionKey);
+            partitionKey,
+            SoftDelete);
     }
 
-    public Task<int> DeleteAsync(Expression<Func<TEntity, bool>> predicate)
+    public async Task<int> DeleteAsync(Expression<Func<TEntity, bool>> predicate)
     {
-        return _database.DeleteAsync(predicate);
+        if (!SoftDeleteState.IsEnabled)
+        {
+            return await _database.DeleteAsync(predicate);
+        }
+
+        var entities = await QueryAsync(predicate);
+        var tasks = new List<Task>();
+        foreach (var entity in entities)
+        {
+            SoftDelete(entity);
+            tasks.Add(_database.ReplaceAsync(entity));
+        }
+
+        await Task.WhenAll(tasks);
+        return entities.Count;
     }
 }
